+++ WorldSvr/OnSkillToAction.c
@@ -12,7 +12,7 @@ CLIENT_PROCEDURE_BINDING(SKILL_TO_ACTION) {
 
     S2C_DATA_SKILL_TO_ACTION* Notification = PacketInit(S2C_DATA_SKILL_TO_ACTION);
     Notification->Command = S2C_SKILL_TO_ACTION;
-    Notification->CharacterIndex = Character->CharacterIndex;
+    Notification->CharacterIndex = (UInt32)Character->CharacterIndex;
     Notification->TargetIndex = Packet->TargetIndex;
     Notification->ActionIndex = Packet->ActionIndex;
     Notification->X = Packet->X;
